pyproject.toml
[tool.poetry]
name = "stutgard"
version = "0.1.0"
description = ""
authors = ["CHAMOMILE <CHAMOM2LES@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
requests = "^2.32.3"
pydantic = "^2.9.2"
celery = "^5.4.0"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"

[tool.poetry.group.dev.dependencies]
djangorestframework = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
